# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API completa para gestão de equinos com Tokenização RWA (Real World Assets). Inclui: autenticação, gestão de equinos, tokenização, leilões, exames laboratoriais, nutrição com IA, treinamento, rankings, financeiro e muito mais.",
        "title": "Equinoid API - Sistema Perfeito",
        "termsOfService": "https://equinoid.com/terms",
        "contact": {
            "name": "Equinoid Support",
            "url": "https://equinoid.com/support",
            "email": "suporte@equinoid.com"
        },
        "license": {
            "name": "Proprietária",
            "url": "https://equinoid.com/license"
        },
        "version": "2.0"
    },
    "host": "seahorse-app-28du8.ondigitalocean.app",
    "basePath": "/api/v1",
    "paths": {
...
        "/equinos/{equinoid}/eventos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna todos os eventos de um equino específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Eventos"
                ],
                "summary": "Listar eventos de um equino",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Equinoid do equino",
                        "name": "equinoid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/eventos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todos os eventos com paginação e filtros",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Eventos"
                ],
                "summary": "Listar eventos",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Itens por página",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por categoria",
                        "name": "categoria",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrar por tipo",
                        "name": "tipo_evento",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Data início",
                        "name": "data_inicio",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Data fim",
                        "name": "data_fim",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria um novo evento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Eventos"
                ],
                "summary": "Criar evento",
                "parameters": [
                    {
                        "description": "Dados do evento",
                        "name": "evento",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateEventoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/eventos/{evento_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os dados de um evento específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Eventos"
                ],
                "summary": "Buscar evento por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do evento",
                        "name": "evento_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza os dados de um evento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Eventos"
                ],
                "summary": "Atualizar evento",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do evento",
                        "name": "evento_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualização",
                        "name": "evento",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove um evento do sistema",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Eventos"
                ],
                "summary": "Deletar evento",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do evento",
                        "name": "evento_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateEventoRequest": {
            "type": "object",
            "required": [
                "data_evento",
                "descricao",
                "tipo_evento"
            ],
            "properties": {
                "aceita_patrocinio": {
                    "type": "boolean"
                },
                "categoria": {
                    "type": "string"
                },
                "data_evento": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "documentos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DocumentoEvento"
                    }
                },
                "informacoes_patrocinio": {
                    "type": "string"
                },
                "local": {
                    "type": "string"
                },
                "nome_evento": {
                    "type": "string"
                },
                "organizador": {
                    "type": "string"
                },
                "participante": {
                    "type": "boolean"
                },
                "particularidades": {
                    "type": "string"
                },
                "resultados": {
                    "type": "string"
                },
                "tipo_evento": {
                    "$ref": "#/definitions/models.TipoEvento"
                },
                "tipo_evento_competitivo": {
                    "type": "string"
                },
                "tipo_evento_publico": {
                    "type": "string"
                },
                "valor_inscricao": {
                    "type": "number"
                },
                "veterinario_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.DocumentoEvento": {
            "type": "object",
            "required": [
                "nome",
                "tipo"
            ],
            "properties": {
                "conteudo": {
                    "description": "Base64 encoded content",
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "tamanho": {
                    "type": "integer"
                },
                "tipo": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "models.TipoEvento": {
            "type": "string",
            "enum": [
                "nascimento",
                "vacina",
                "exame",
                "cirurgia",
                "transferencia",
                "obito",
                "competicao",
                "reproducao",
                "treinamento",
                "viagem",
                "certificacao"
            ],
            "x-enum-varnames": [
                "TipoEventoNascimento",
                "TipoEventoVacina",
                "TipoEventoExame",
                "TipoEventoCirurgia",
                "TipoEventoTransferencia",
                "TipoEventoObito",
                "TipoEventoCompeticao",
                "TipoEventoReproducao",
                "TipoEventoTreinamento",
                "TipoEventoViagem",
                "TipoEventoCertificacao"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Formato: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}